+++ typescript/financial-dashboard/src/app/layout-client.tsx
@@ -1,8 +1,13 @@
 "use client";
 
 import { ReactNode } from "react";
+import Link from "next/link";
 
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
+import { HomeIcon } from "lucide-react";
+
+import { AiCopilotButton } from "@/components/molecules/ai-copilot/AiCopilotButton";
+import { AiCopilotProvider } from "@/components/molecules/ai-copilot/AiCopilotProvider";
 
 const queryClient = new QueryClient();
 
@@ -12,6 +17,36 @@ type Props = {
 
 export default function RootLayoutClient({ children }: Props) {
   return (
-    <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
+    <AiCopilotProvider>
+      <header className="sticky top-0 z-10 flex h-16 items-center gap-4 border-b bg-background px-4 md:px-6">
+        <nav className="flex flex-col gap-6 text-lg font-medium md:flex-row md:items-center md:gap-5 md:text-sm lg:gap-6">
+          <Link
+            href="/"
+            className="flex items-center gap-2 text-lg font-semibold md:text-base"
+          >
+            <HomeIcon className="size-5" />
+            <h1>GPU Shop</h1>
+          </Link>
+          <Link
+            href="/products"
+            className="text-muted-foreground transition-colors hover:text-foreground"
+          >
+            Products
+          </Link>
+          <Link
+            href="/users"
+            className="text-muted-foreground transition-colors hover:text-foreground"
+          >
+            Users
+          </Link>
+        </nav>
+
+        <span className="flex-1"></span>
+
+        <AiCopilotButton />
+      </header>
+
+      <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
+    </AiCopilotProvider>
   );
 }
