+++ typescript/financial-dashboard/src/components/molecules/ai-copilot/AiCopilotProvider.tsx
@@ -5,6 +5,8 @@ type AiCopilotContextValue = {
   setContext: (context: unknown) => void;
   open: boolean;
   setOpen: (open: boolean) => void;
+  active: boolean;
+  setActive: (active: boolean) => void;
 };
 
 const AiCopilotContext = createContext<AiCopilotContextValue | null>(null);
@@ -14,10 +16,11 @@ export const AiCopilotProvider: React.FC<{ children: React.ReactNode }> = ({
 }) => {
   const [context, setContext] = useState<unknown>();
   const [open, setOpen] = useState(false);
+  const [active, setActive] = useState(false);
 
   const contextValue = useMemo<AiCopilotContextValue>(
-    () => ({ context, setContext, open, setOpen }),
-    [JSON.stringify(context), open]
+    () => ({ context, setContext, open, setOpen, active, setActive }),
+    [JSON.stringify(context), open, active]
   );
 
   return (
